+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 ### Fixed
 - show right default background for theme in background preview in settings
 - handle invalid theme metadata better (don't display no themes anymore when one has invalid metadata)
+- Fix copy of labeled links
 - Fix hard to alter text in the middle of "Video chat instance" input field (see #2016)
 
 ### Added
 - open message info when clicking on the error status icon of a message
 - add deltachat desktop version & build info to the logfiles
+- Add warn dialog for urls containing punycode
 - add a minimalistic theme featuring an irc like message view
 
 ### Changed